% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFuncs.R
\name{getDescMatrix}
\alias{getDescMatrix}
\alias{getAncVec}
\alias{getNodesOfDepth}
\title{Get descendants of nodes in matrix}
\usage{
getDescMatrix(tree)

getAncVec(tree)

getNodesOfDepth(mat, root, n)
}
\arguments{
\item{tree}{Rooted binary tree of class "phylo", >= 2 leaves
(no specific node enumeration order required)}

\item{mat}{descendants matrix from \code{getDescMatrix}}

\item{root}{Number (label) of the root of the tree}

\item{n}{Number of leaves of the tree}
}
\value{
numeric matrix  (2n-1 rows, 2 columns)

numeric vector  (size 2n-1)

numeric matrix  (n rows, n columns)
}
\description{
\code{getDescMatrix} - Determines a matrix that contains the descendants of
node i in row i.

\code{getAncVec} - Determines a vector that contains the ancestor of
node i in position i.

\code{getAncVec} - Determines a matrix that contains the nodes of
depth i in row i.
}
\examples{
mat <- cbind(c(7,7,6,5,5,6),c(1,2,3,4,6,7))
tree <- list(edge=mat, tip.label=c("","","",""), Nnode=3)
attr(tree, "class") <- "phylo"
getDescMatrix(tree)
getAncVec(tree)
getNodesOfDepth(mat=getDescMatrix(tree),
root=which(is.na(getAncVec(tree))), n=tree$Nnode+1)
}
\author{
Sophie Kersting
}
