% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modCherryIndex.R
\name{modCherryIndex}
\alias{modCherryIndex}
\title{Modified cherry index of a rooted binary tree}
\usage{
modCherryIndex(tree)
}
\arguments{
\item{tree}{Rooted binary tree of class "phylo"
(no specific node enumeration order required)}
}
\value{
the number of leaves not in a cherry (numeric)
}
\description{
\code{modCherryIndex} - This function calculates the modified cherry index
value, i.e. the number of leaves that are not in a cherry, for rooted
binary trees.
}
\examples{
mat <- cbind(c(7,7,6,5,5,6),c(1,2,3,4,6,7))
tree <- list(edge=mat, tip.label=c("","","",""), Nnode=3)
attr(tree, "class") <- "phylo"
modCherryIndex(tree)
}
\author{
Sophie Kersting
}
