% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogersJ.R
\name{rogersJ}
\alias{rogersJ}
\title{Rogers J Index of a rooted binary tree}
\usage{
rogersJ(tree)
}
\arguments{
\item{tree}{Rooted binary tree of class "phylo"
(no specific node enumeration order required)}
}
\value{
the number of unbalanced nodes (numeric)
}
\description{
\code{rogersJ} - This function calculates the Rogers J index value, the
number of interior nodes that are not balanced (i.e. their two pending
subtrees differ in size), for rooted binary trees.
}
\examples{
mat <- cbind(c(7,7,6,5,5,6),c(1,2,3,4,6,7))
tree <- list(edge=mat, tip.label=c("","","",""), Nnode=3)
attr(tree, "class") <- "phylo"
rogersJ(tree)
}
\author{
Sophie Kersting
}
