% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmeTree.R
\docType{package}
\name{symmeTree}
\alias{symmeTree}
\title{symmeTree : Calculate the symmetry nodes, the modified cherry and the Rogers
J Index of rooted binary trees.}
\description{
Calculate the symmetry nodes index, the modified cherry index or the Rogers J
index of rooted binary trees in the phylo format. Furthermore, calculate the
binary weight and the minimal value of the symmetry nodes and J index.
}
