% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symNodesIndex.R
\name{symNodesIndex}
\alias{symNodesIndex}
\title{Symmetry Nodes Index of a rooted binary tree}
\usage{
symNodesIndex(tree)
}
\arguments{
\item{tree}{Rooted binary tree of class "phylo"
(no specific node enumeration order required)}
}
\value{
number of interior nodes that are not symmetry nodes (numeric)
}
\description{
\code{symNodesIndex} - This function calculates the symmetry nodes index
value, i.e. the number of interior nodes that are not symmetry nodes, for
rooted binary trees.
}
\examples{
mat <- cbind(c(7,7,6,5,5,6),c(1,2,3,4,6,7))
tree <- list(edge=mat, tip.label=c("","","",""), Nnode=3)
attr(tree, "class") <- "phylo"
symNodesIndex(tree)
}
\author{
Sophie Kersting
}
