% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFuncs.R
\name{getDescMatrix}
\alias{getDescMatrix}
\alias{getAncVec}
\alias{getNodesOfDepth}
\alias{myBucketLexicoSort}
\title{Auxiliary functions for handling nodes in trees}
\usage{
getDescMatrix(tree)

getAncVec(tree)

getNodesOfDepth(mat, root, n)

myBucketLexicoSort(workLabs)
}
\arguments{
\item{tree}{Rooted binary tree of class "phylo", >= 2 leaves
(no specific node enumeration order required)}

\item{mat}{Descendants matrix from \code{getDescMatrix}}

\item{root}{Number (label) of the root of the tree}

\item{n}{Number of leaves of the tree}

\item{workLabs}{numeric matrix (2 columns)}
}
\value{
\code{desc_mat} numeric matrix  (2n-1 rows, 2 columns)

\code{anc_vec} numeric vector  (size 2n-1)

\code{nodes_of_depth} numeric matrix  (n rows, n columns)

\code{ranking} numeric vector
}
\description{
\code{getDescMatrix} - Determines a matrix that contains the descendants of
node i in row i.

\code{getAncVec} - Determines a vector that contains the ancestor of
node i in position i.

\code{getNodesOfDepth} - Determines a matrix that contains the nodes of
depth i in row i.

\code{myBucketLexicoSort} - Sorts the pairs of numbers lexicographically and
returns ranking. Uses bucket sort.
}
\examples{
mat <- cbind(c(7,7,6,5,5,6),c(1,2,3,4,6,7))
tree <- list(edge=mat, tip.label=c("","","",""), Nnode=3)
attr(tree, "class") <- "phylo"
getDescMatrix(tree)
getAncVec(tree)
getNodesOfDepth(mat=getDescMatrix(tree),
root=which(is.na(getAncVec(tree))), n=tree$Nnode+1)
myWorkLabs <- cbind(c(0,1,2,3,1,0),c(0,2,2,4,1,0))
myBucketLexicoSort(myWorkLabs)
}
\author{
Sophie Kersting
}
